requirements-macos.txt
# macOS-specific requirements for TinyRecursiveModels
# This replaces triton with MLX dependencies

torch
einops
tqdm
coolname
pydantic
argdantic
wandb
omegaconf
hydra-core
huggingface_hub
packaging 
ninja 
wheel 
setuptools 
setuptools-scm
pydantic-core
numba

# MLX dependencies for macOS (replaces triton)
mlx
mlx-lm
transformers
tokenizers
safetensors
regex

# Note: adam-atan2 is replaced with adam_atan2_macos.py (included in repo)
# This avoids CUDA dependency issues on macOS
